% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephant.R
\name{:=}
\alias{:=}
\title{Create elephant object}
\usage{
`:=`(x, value)
}
\arguments{
\item{x}{symbol to assign to}

\item{value}{value to assign to x}
}
\description{
\verb{:=} works like a regular assignment but creates an elephant object, which is
like a regular object but with a memory of how it was created, this memory is
recursive if variables used in the \code{value} expression are elephants themselves.
}
