% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Math.elephant}
\alias{Math.elephant}
\alias{Ops.elephant}
\alias{Complex.elephant}
\alias{Summary.elephant}
\alias{+.elephant}
\alias{-.elephant}
\alias{*.elephant}
\alias{/.elephant}
\alias{^.elephant}
\alias{\%\%.elephant}
\alias{\%/\%.elephant}
\title{methods for class 'elephant'}
\usage{
\method{Math}{elephant}(x, ...)

\method{Ops}{elephant}(e1, e2)

\method{Complex}{elephant}(z)

\method{Summary}{elephant}(x, ..., na.rm = FALSE)

\method{+}{elephant}(e1, e2)

\method{-}{elephant}(e1, e2)

\method{*}{elephant}(e1, e2)

\method{/}{elephant}(e1, e2)

\method{^}{elephant}(e1, e2)

\method{\%\%}{elephant}(e1, e2)

\method{\%/\%}{elephant}(e1, e2)
}
\arguments{
\item{x}{objects.}

\item{...}{further arguments passed to methods.}

\item{e1}{objects.}

\item{e2}{objects.}

\item{z}{objects.}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
methods for class 'elephant'
}
