% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{Math.elephant}
\alias{Math.elephant}
\alias{Ops.elephant}
\alias{Complex.elephant}
\alias{Summary.elephant}
\alias{+.elephant}
\alias{-.elephant}
\alias{*.elephant}
\alias{/.elephant}
\alias{^.elephant}
\alias{\%\%.elephant}
\alias{\%/\%.elephant}
\title{methods for class 'elephant'}
\usage{
\method{Math}{elephant}(x, ...)

\method{Ops}{elephant}(x, ...)

\method{Complex}{elephant}(x, ...)

\method{Summary}{elephant}(x, ...)

\method{+}{elephant}(e1, e2)

\method{-}{elephant}(e1, e2)

\method{*}{elephant}(e1, e2)

\method{/}{elephant}(e1, e2)

\method{^}{elephant}(e1, e2)

\method{\%\%}{elephant}(e1, e2)

\method{\%/\%}{elephant}(e1, e2)
}
\arguments{
\item{x}{object}

\item{...}{passed to other methods}

\item{e1}{object}

\item{e2}{object}
}
\description{
methods for class 'elephant'
}
