% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elephant.R
\name{print.elephant}
\alias{print.elephant}
\title{Print elephant object}
\usage{
\method{print}{elephant}(x, ..., print.value = TRUE)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}

\item{print.value}{if \code{FALSE} only the elephant's memory is printed.}
}
\description{
Print elephant object.
}
